First part:
PersistenceManager file is created. This file connects default memory and Data that should be saved in there and retrieved from there.

This data is the login and avatar image of the the GitHub user who was favorited.

Function ‘retrieveFavorites’ add a new user to [Follower] array or throws an error of GFError type. This function decodes the data fetched from defaults.

Function ‘save’ encodes it back to defaults.

These two functions are the elements of one static function ‘updateWith’.

2. When ‘updateWith’ is called, ‘PersistenceActionType’ is being checked (either it is .add or .remove).

If ‘retrievedFavorites’ array does not contain a new favorited user & ‘.add’ actionType is swiped  a new user Data will be appended to the  ‘retrievedFavorites’ array.

If ‘retrievedFavorites’ array contain a new favorited user & ‘.remove’ actionType is swiped, then the data about this user will be deleted from ‘retrievedFavorites’ array.

Second part:

In ‘FollowerListVC’ there is @objc func ‘addButtonTapped’. Inside that function function ‘getUserInfo’ is being called.

If call is successful, ‘PersistenceManager.updateWith’ fucntion is called. ‘.add’ parameter is passed in there. If there is no error, a new user Data will be appended to the  ‘retrievedFavorites’ array and message about successful favoriting the  user will be sent to the screen.

If call is unsuccessful, message about unsuccessful favoriting the user will be sent to the screen.

Third part:

‘FavoriteCell’ is created programmatically. Inside there is a function ‘setTextAndImageFor’ that is created for downloading the image for the ‘avatarImageView’ and text for the ‘userNameLabel’ as 'login name.