1.In ‘FavoritesListVC’ UITableView() was initialized. ‘UITableViewDataSource’ was initialized.
numberOfRowsInSection was set to the number of users in ‘favorites’ array with the method ‘favorites.count’.

In another method ‘cellForRowAt’ a reusable cell was created. every time we scroll down, cellForRowAt is called, and 'setTextAndImageFor' is called through ‘FavoriteCell’. And in 'let favorite = favorites[indexPath.row]’ new cells are being created as many as many users in ‘favorites’ array.

2.Delegate methods for reusable cells were initialized.

In function ‘didSelectRowAt’ - ‘FollowerListVC’ is pushed every time when a user is tapping on particular row with a particular indexPath.

In function 'commit editingStyle’. When gesture is to ‘.delete’ the cell, then the favorited user is being ‘.remove’ from the ‘favorites’ array, and the row with that user info is being deleted from the ‘tableView’. Then, through ‘PersistenceManager’ initializer we choose actionType ‘.remove’, and this removes user’s data from the defaults.

3.In function ‘getFavorites’ we retrieveFavorites from defaults. If ‘favorites’ array is ‘.isEmpty’, ‘showEmptyStateView’ is being showed. If retrieving the Favorites is successful, then ‘tableView’ is being called and being ‘.reloadData’ -> through this function we populate 'var favorites: [Follower]   = []’ array, and then create as many cells as we need it’s said in point 1 (‘cellForRowAt’).